Restructure environment of development
