--- .eslintignore
@@ -1,3 +1,0 @@
-.eslintrc.json
-scripts/release.js
-/playwright-report
