+++ package.json
@@ -1,21 +1,26 @@
 {
-  "name": "GIKYOKU",
+  "name": "gikyoku",
   "version": "1.0.0",
-  "description": "implement the screenplay pattern for Playwright.",
+  "description": "Implement the screenplay pattern for Playwright.",
   "main": "index.js",
   "types": "index.d.ts",
   "commitHooks": false,
   "scripts": {
-    "build": "rm -r lib || true && tsc --project ./tsconfig-build.json",
-    "test": "npx playwright test __tests__ --config=__tests__/playwright.config.ts",
-    "test:api": "npx playwright test __tests__/api.spec.ts --config=__tests__/playwright.config.ts",
-    "test:web": "npx playwright test __tests__/web.spec.ts --config=__tests__/playwright.config.ts",
-    "lint:ts": "tsc --noEmit",
-    "lint:js": "eslint . --cache",
-    "lint": "npm run lint:js && npm run lint:ts",
-    "prerelease": "npm run lint && npm run build && npm run test",
-    "release": "node scripts/release.js",
-    "preversion": "npm run lint"
+    "prepare": "run-p prepare:*",
+    "prepare:husky": "husky install",
+    "prepare:playwright": "playwright install",
+    "prebuild": "failsafe clean lint format test",
+    "build": "tsc --project ./tsconfig-build.json",
+    "test": "npx playwright test",
+    "test:api": "npx playwright test tests/api",
+    "test:web": "npx playwright test tests/web",
+    "lint": "eslint ./ --fix",
+    "format": "prettier ./src/**/*.ts --write",
+    "clean": "rimraf dist && rimraf report && rimraf result",
+    "postversion": "git push && git push --tags",
+    "package": "node scripts/package.js",
+    "prerelease": "npm run build && npm run package",
+    "release": "npm publish --access=public ./dist"
   },
   "repository": {
     "type": "git",
@@ -33,19 +38,22 @@
   },
   "homepage": "https://github.com/aYukiYoshida/GIKYOKU#readme",
   "devDependencies": {
-    "@types/node": "^17.0.23",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "eslint": "^8.12.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.4",
-    "typescript": "^4.6.3",
-    "v8-to-istanbul": "^9.0.1"
+    "@types/node": "^17.0.45",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint": "^8.51.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "husky": "^8.0.3",
+    "lint-staged": "^14.0.1",
+    "npm-failsafe": "^1.1.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^3.0.3",
+    "rimraf": "^5.0.5",
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
-    "@playwright/test": "^1.34.0"
+    "@playwright/test": "^1.39.0"
   },
   "dependencies": {
     "@testla/screenplay": "^0.3.0"
@@ -55,5 +63,11 @@
   },
   "publishConfig": {
     "access": "public"
+  },
+  "lint-staged": {
+    "*.ts": [
+      "eslint --fix",
+      "prettier --write"
+    ]
   }
 }
