+++ tsconfig-build.json
@@ -1,13 +1,9 @@
 {
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "rootDir": "./src",
-    },
-    "include": [
-        "src"
-      ],
-      "exclude": [
-        "node_modules",
-        "__tests__"
-      ]
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+  },
+  "include": [
+    "src/**/*"
+  ]
 }
